requirements.txt
requests~=2.27.1
pyftpdlib~=1.5.7
PyYAML~=6.0